% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{anno_vertex}
\alias{anno_vertex}
\alias{anno_node}
\title{Use data.frame to annotate vertexes of an igraph}
\usage{
anno_vertex(go, anno_tab)
}
\arguments{
\item{go}{metanet an igraph object}

\item{anno_tab}{a dataframe using to annotate (with rowname or a name column)}
}
\value{
a annotated igraph object
}
\description{
Use data.frame to annotate vertexes of an igraph
}
\examples{
data("c_net")
data("otutab", package = "pcutils")
anno_vertex(co_net, taxonomy)
}
