% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4.other_plot.R
\name{as.ggig}
\alias{as.ggig}
\title{Transfer an igraph object to a ggig}
\usage{
as.ggig(go, coors = NULL)
}
\arguments{
\item{go}{igraph or meatnet}

\item{coors}{coordinates for nodes,columns: name, X, Y}
}
\value{
ggig object
}
\description{
Transfer an igraph object to a ggig
}
\examples{
as.ggig(co_net, coors = c_net_layout(co_net)) -> ggig
plot(ggig)
as.ggig(multi1, coors = c_net_layout(multi1)) -> ggig
plot(ggig, labels_num = 0.3)
}
\seealso{
Other plot: 
\code{\link{c_net_plot}()},
\code{\link{input_gephi}()},
\code{\link{netD3plot}()},
\code{\link{olympic_rings_net}()},
\code{\link{plot.ggig}()},
\code{\link{twocol_edgelist}()},
\code{\link{venn_net}()}
}
\concept{plot}
