% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_arc}
\alias{as_arc}
\title{Layout as a arc}
\usage{
as_arc(angle = 0, arc = pi)
}
\arguments{
\item{angle}{anticlockwise rotation angle}

\item{arc}{the radian of arc}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a arc
}
\examples{
as_arc()(co_net)
c_net_plot(co_net, coors = as_arc(pi / 2), rescale = FALSE)
}
\seealso{
Other layout: 
\code{\link{as_circle_tree}()},
\code{\link{as_line}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
