% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_circle_tree}
\alias{as_circle_tree}
\title{Layout as a circle_tree}
\usage{
as_circle_tree()
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a circle_tree
}
\examples{
as_circle_tree()(co_net)
}
