% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{as_coors}
\alias{as_coors}
\title{Transfer to a coors object}
\usage{
as_coors(coors, curved = NULL)
}
\arguments{
\item{coors}{data.frame}

\item{curved}{line curved}
}
\value{
coors object
}
\description{
Transfer to a coors object
}
