% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{as_line}
\alias{as_line}
\title{Layout as a line}
\usage{
as_line(angle = 0)
}
\arguments{
\item{angle}{anticlockwise rotation angle}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a line
}
\examples{
as_line()(co_net)
c_net_plot(co_net, coors = as_line(pi / 2))
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_circle_tree}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
