% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{as_multi_layer}
\alias{as_multi_layer}
\title{Layout as a multi_layer}
\usage{
as_multi_layer(n = 3, layout = on_grid())
}
\arguments{
\item{n}{how many arcs of this multi_layer}

\item{layout}{see method in \code{\link[MetaNet]{c_net_layout}}}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a multi_layer
}
\examples{
as_multi_layer()(co_net)
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_circle_tree}()},
\code{\link{as_line}()},
\code{\link{as_poly_sector}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
