% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{as_poly_sector}
\alias{as_poly_sector}
\title{Layout as a multi_layer}
\usage{
as_poly_sector(n = 3)
}
\arguments{
\item{n}{how many arcs of this multi_layer}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a multi_layer
}
\examples{
as_poly_sector()(co_net)
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_circle_tree}()},
\code{\link{as_line}()},
\code{\link{as_multi_layer}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
