% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_polycircle}
\alias{as_polycircle}
\title{Layout as a polycircle}
\usage{
as_polycircle(n = 2)
}
\arguments{
\item{n}{how many circles of this polycircle}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a polycircle
}
\examples{
as_polycircle()(co_net)
}
