% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_polygon}
\alias{as_polygon}
\title{Layout as a polygon}
\usage{
as_polygon(n = 3)
}
\arguments{
\item{n}{how many edges of this polygon}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a polygon
}
\examples{
as_polygon()(co_net)
}
