% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_build}
\alias{c_net_build}
\title{Construct a network for table corr}
\usage{
c_net_build(corr, r_thres = 0.6, p_thres = 0.05, del_single = T)
}
\arguments{
\item{corr}{result from c_net_cal()}

\item{r_thres}{r_threshold (default:>0.6)}

\item{p_thres}{p_threshold (default:<0.05)}

\item{del_single}{should delete single vertexes?}
}
\value{
an igraph object
}
\description{
Construct a network for table corr
}
\examples{
data("otutab")
t(otutab) -> totu
c_net_cal(totu) -> corr
c_net_build(corr) -> co_net

}
