% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{c_net_difference}
\alias{c_net_difference}
\title{Difference two networks}
\usage{
c_net_difference(go1, go2, ...)
}
\arguments{
\item{go1}{metanet object}

\item{go2}{metanet object}

\item{...}{add}
}
\value{
metanet
}
\description{
Difference two networks
}
