% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_filter}
\alias{c_net_filter}
\title{Filter a network according to some attributes}
\usage{
c_net_filter(go, ..., mode = "v")
}
\arguments{
\item{go}{metanet object}

\item{...}{some attributes of vertex and edge}

\item{mode}{"v" or "e"}
}
\value{
metanet
}
\description{
Filter a network according to some attributes
}
\examples{
data("multi_net")
c_net_filter(multi1, v_group \%in\% c("omic1", "omic2"))
}
\seealso{
Other manipulate: 
\code{\link{anno_edge}()},
\code{\link{anno_vertex}()},
\code{\link{c_net_annotate}()},
\code{\link{c_net_save}()},
\code{\link{c_net_union}()},
\code{\link{get_e}()},
\code{\link{get_n}()},
\code{\link{get_v}()},
\code{\link{is_metanet}()}
}
\concept{manipulate}
