% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_from_edgelist}
\alias{c_net_from_edgelist}
\title{Construct a network from Edge_list dataframe}
\usage{
c_net_from_edgelist(edgelist, direct = F, e_type = NULL, e_class = NULL)
}
\arguments{
\item{edgelist}{first is source, second is target, others are annotation}

\item{direct}{logical}

\item{e_type}{set e_type}

\item{e_class}{set e_class}
}
\description{
Construct a network from Edge_list dataframe
}
