% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{c_net_intersect}
\alias{c_net_intersect}
\title{Intersect two networks}
\usage{
c_net_intersect(go1, go2, ...)
}
\arguments{
\item{go1}{metanet object}

\item{go2}{metanet object}

\item{...}{add}
}
\value{
metanet
}
\description{
Intersect two networks
}
