% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_load}
\alias{c_net_load}
\title{Load network file}
\usage{
c_net_load(filename, format = "data.frame")
}
\arguments{
\item{filename}{filename}

\item{format}{"data.frame","graphml"}
}
\value{
metanet
}
\description{
Load network file
}
\seealso{
Other manipulate: 
\code{\link{anno_edge}()},
\code{\link{anno_vertex}()},
\code{\link{c_net_annotate}()},
\code{\link{c_net_filter}()},
\code{\link{c_net_save}()},
\code{\link{c_net_union}()},
\code{\link{get_e}()},
\code{\link{get_n}()},
\code{\link{get_v}()},
\code{\link{is_metanet}()}
}
\concept{manipulate}
