% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{c_net_plot}
\alias{c_net_plot}
\title{Plot a co-network}
\usage{
c_net_plot(go, coors = NULL, ...)
}
\arguments{
\item{go}{a igraph object}

\item{coors}{the coordinates you saved}

\item{...}{additional parameters for plot.igraph()}
}
\value{
a network plot
}
\description{
Plot a co-network
}
\examples{
data("c_net")
c_net_plot(co_net)
c_net_plot(co_net2)
c_net_plot(co_net3)
}
