% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_save}
\alias{c_net_save}
\title{Save network file}
\usage{
c_net_save(go, filename = "net", format = "data.frame")
}
\arguments{
\item{go}{metanet network}

\item{filename}{filename}

\item{format}{"data.frame","graphml"}
}
\description{
Save network file
}
