% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{cbind_new}
\alias{cbind_new}
\title{bind two df with same columns, the last df will replace first df}
\usage{
cbind_new(df, df1)
}
\description{
bind two df with same columns, the last df will replace first df
}
