% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{check_tabs}
\alias{check_tabs}
\title{Check tables and extract common samples}
\usage{
check_tabs(...)
}
\arguments{
\item{...}{tables}
}
\value{
formatted tables
}
\description{
Check tables and extract common samples
}
\examples{
data("otutab", package = "pcutils")
check_tabs(otutab)
}
