% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{clean_igraph}
\alias{clean_igraph}
\title{Clean a igraph object}
\usage{
clean_igraph(go, direct = TRUE)
}
\arguments{
\item{go}{igraph, metanet objects}

\item{direct}{direct?}
}
\value{
a igraph object
}
\description{
Clean a igraph object
}
