% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{clean_multi_edge_metanet}
\alias{clean_multi_edge_metanet}
\title{Clean multi edge metanet to plot}
\usage{
clean_multi_edge_metanet(go)
}
\arguments{
\item{go}{metanet object}
}
\value{
metanet object
}
\description{
Clean multi edge metanet to plot
}
\examples{
g <- igraph::make_ring(2)
g <- igraph::add.edges(g, c(1, 1, 1, 1, 2, 1))
plot(g)
plot(clean_multi_edge_metanet(g))
}
