% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{rand_net_par}
\alias{rand_net_par}
\alias{compare_rand}
\title{Net_pars of many random network}
\usage{
rand_net_par(go, reps = 99, threads = 1)

compare_rand(pars, randp, index = c("ave_path_len", "clusteringC"))
}
\arguments{
\item{go}{igraph}

\item{reps}{simulation time}

\item{threads}{threads}

\item{pars}{your net pars resulted by net_pars()}

\item{randp}{random networks pars resulted by rand_net_par()}

\item{index}{compared indexes: "ave_path_len","clusteringC" or else}
}
\value{
ggplot
}
\description{
Net_pars of many random network

Compare some indexes between your net with random networks
}
\examples{
\donttest{
\dontrun{
rand_net_par(co_net_rmt,reps = 30)->randp
net_par(co_net_rmt,fast = FALSE)->pars
compare_rand(pars,randp)
}}
}
