% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{df2net_tree}
\alias{df2net_tree}
\title{Transfer a dataframe to a network edgelist.}
\usage{
df2net_tree(test, fun = sum)
}
\arguments{
\item{test}{df}

\item{fun}{default: sum}
}
\value{
metanet
}
\description{
Transfer a dataframe to a network edgelist.
}
\examples{
data("otutab", package = "pcutils")
cbind(taxonomy, num = rowSums(otutab))[1:20, ] -> test
df2net_tree(test) -> ttt
}
