% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.topological.R
\name{extract_sample_net}
\alias{extract_sample_net}
\title{Extract each sample network from the whole network}
\usage{
extract_sample_net(
  whole_net,
  otutab,
  threads = 1,
  save_net = FALSE,
  fast = TRUE,
  remove_negative = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{whole_net}{the whole network}

\item{otutab}{otutab, columns are samples, these columns will be extract}

\item{threads}{threads, default: 1}

\item{save_net}{should save these sub_nets? FALSE or a filename}

\item{fast}{less indexes for faster calculate ?}

\item{remove_negative}{remove negative edge or not? default: FALSE}

\item{verbose}{verbose}
}
\value{
a dataframe contains all sub_net parameters
}
\description{
Extract each sample network from the whole network
}
\examples{
data(otutab, package = "pcutils")
extract_sample_net(co_net, otutab) -> sub_net_pars
}
\seealso{
Other topological: 
\code{\link{fit_power}()},
\code{\link{get_group_skeleton}()},
\code{\link{links_stat}()},
\code{\link{nc}()},
\code{\link{net_par}()},
\code{\link{rand_net_par}()},
\code{\link{rand_net}()},
\code{\link{smallworldness}()}
}
\concept{topological}
