% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{fit_power}
\alias{fit_power}
\title{Fit power-law distribution for a igraph}
\usage{
fit_power(go, pvalue = F)
}
\arguments{
\item{go}{igraph}
}
\value{
ggplot
}
\description{
Fit power-law distribution for a igraph
}
\examples{
fit_power(co_net)
}
