% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{g_lay}
\alias{g_lay}
\alias{g_lay_nice}
\title{Layout with group}
\usage{
g_lay(
  go,
  group = "module",
  zoom1 = 20,
  zoom2 = 3,
  layout1 = in_circle(),
  layout2 = in_circle()
)

g_lay_nice(go, group = "module")
}
\arguments{
\item{go}{igraph object}

\item{group}{group name (default:module)}

\item{zoom1}{big network layout size}

\item{zoom2}{average sub_network layout size}

\item{layout1}{layout1 method, one of
{1.numeric: use adjusted circle}
{2.a dataframe: rowname is group, two columns are X and Y}
{3.function: \code{\link[igraph]{layout_}} default: in_circle()}}

\item{layout2}{one of \code{\link[igraph]{layout_}}}
}
\value{
coors
}
\description{
Layout with group

Layout with group nicely
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
g_lay(co_net_modu,group ="module" ,zoom2 = 5,layout2 =nicely())->oridata
modu_plot(co_net_modu,coors = oridata)
g_lay_nice(co_net_modu,group ="module")->oridata
modu_plot(co_net_modu,coors = oridata)

}
