% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{g_layout_poly_sector}
\alias{g_layout_poly_sector}
\title{Layout with group}
\usage{
g_layout_poly_sector(
  go,
  group = "v_group",
  group_order = NULL,
  group2 = NULL,
  group2_order = NULL,
  space = pi/4,
  jitter = 0,
  scale_node_num = TRUE,
  type = c("regular", "random"),
  mode = c("surface", "boundary"),
  curved = NULL
)
}
\arguments{
\item{go}{igraph}

\item{group}{group name (default:v_group)}

\item{group_order}{group_order}

\item{group2}{group2 name, will order nodes in each group according to group2_order}

\item{group2_order}{group2_order}

\item{space}{the space between each arc, default: pi/4}

\item{jitter}{for surface-regular, defalut 0}

\item{scale_node_num}{scale with the node number in each group}

\item{type}{Type of distribution: "random", "regular"}

\item{mode}{"surface", "boundary"}

\item{curved}{Optional curved attribute for coors}
}
\value{
coors
}
\description{
Layout with group
}
\seealso{
Other g_layout: 
\code{\link{g_layout}()},
\code{\link{g_layout_nice}()},
\code{\link{g_layout_polygon}()}
}
\concept{g_layout}
