% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{get_group_skeleton}
\alias{get_group_skeleton}
\alias{c_net_skeleton}
\alias{skeleton_plot}
\title{Get skeleton network according to a group}
\usage{
get_group_skeleton(go, Group = "v_class", count = NULL, top_N = 8)

skeleton_plot(ske_net, split_e_type = TRUE, ...)
}
\arguments{
\item{go}{network}

\item{Group}{vertex column name}

\item{count}{take which column count, default: NULL}

\item{top_N}{top_N}

\item{ske_net}{skeleton}

\item{split_e_type}{split by e_type? default: TRUE}

\item{...}{additional parameters for \code{\link[igraph]{igraph.plotting}}}
}
\value{
skeleton network
}
\description{
Get skeleton network according to a group

Skeleton plot
}
\examples{
get_group_skeleton(co_net) -> ske_net
skeleton_plot(ske_net)
}
\seealso{
Other topological: 
\code{\link{extract_sample_net}()},
\code{\link{fit_power}()},
\code{\link{links_stat}()},
\code{\link{nc}()},
\code{\link{net_par}()},
\code{\link{rand_net_par}()},
\code{\link{rand_net}()},
\code{\link{smallworldness}()}
}
\concept{topological}
