% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_module_eigen}
\alias{get_module_eigen}
\title{Get module_eigen}
\usage{
get_module_eigen(go_m)
}
\arguments{
\item{go_m}{module metanet}
}
\value{
module_eigen
}
\description{
Get module_eigen
}
\seealso{
Other module: 
\code{\link{filter_n_module}()},
\code{\link{get_community}()},
\code{\link{get_module}()},
\code{\link{module_detect}()},
\code{\link{module_eigen}()},
\code{\link{module_net}()},
\code{\link{summary_module}()},
\code{\link{to_module_net}()},
\code{\link{zp_analyse}()}
}
\concept{module}
