% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_n}
\alias{get_n}
\title{Get network information}
\usage{
get_n(go, simple = FALSE)
}
\arguments{
\item{go}{metanet}

\item{simple}{logical, get simple index}
}
\value{
data.frame
}
\description{
Get network information
}
