% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{get_v}
\alias{get_v}
\title{Get vertex information}
\usage{
get_v(go, name = NULL)
}
\arguments{
\item{go}{metanet object}

\item{name}{attribute name, default: NULL}
}
\value{
data.frame
}
\description{
Get vertex information
}
\seealso{
Other manipulate: 
\code{\link{anno_edge}()},
\code{\link{anno_vertex}()},
\code{\link{c_net_annotate}()},
\code{\link{c_net_filter}()},
\code{\link{c_net_load}()},
\code{\link{c_net_save}()},
\code{\link{c_net_union}()},
\code{\link{get_e}()},
\code{\link{get_n}()},
\code{\link{is_metanet}()}
}
\concept{manipulate}
