% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{input_corr}
\alias{input_corr}
\title{Import corr from .csv file}
\usage{
input_corr(filename)
}
\arguments{
\item{filename}{filename of .corr}
}
\value{
a list with 7 elements:
\item{r}{default: spearman correlation}
\item{p.value}{default: p-value of spearman correlation}
\item{p.adjust}{default p.adjust.method = NULL}
}
\description{
Import corr from .csv file
}
\examples{
\dontrun{
\donttest{
 input_corr("test")->corr
}}
}
