% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4.other_plot.R
\name{input_cytoscape}
\alias{input_cytoscape}
\title{Input a cyjs file exported by Cytoscape}
\usage{
input_cytoscape(file)
}
\arguments{
\item{file}{cyjs file exported by Cytoscape}
}
\value{
list contains the igraph object and coordinates
}
\description{
Input a cyjs file exported by Cytoscape
}
\seealso{
Other plot: 
\code{\link{as.ggig}()},
\code{\link{c_net_plot}()},
\code{\link{input_gephi}()},
\code{\link{netD3plot}()},
\code{\link{olympic_rings_net}()},
\code{\link{plot.ggig}()},
\code{\link{twocol_edgelist}()},
\code{\link{venn_net}()}
}
\concept{plot}
