% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4.other_plot.R
\name{input_gephi}
\alias{input_gephi}
\title{Input a graphml file exported by Gephi}
\usage{
input_gephi(file)
}
\arguments{
\item{file}{graphml file exported by Gephi}
}
\value{
list contains the igraph object and coordinates
}
\description{
Input a graphml file exported by Gephi
}
\seealso{
Other plot: 
\code{\link{as.ggig}()},
\code{\link{c_net_plot}()},
\code{\link{netD3plot}()},
\code{\link{olympic_rings_net}()},
\code{\link{plot.ggig}()},
\code{\link{twocol_edgelist}()},
\code{\link{venn_net}()}
}
\concept{plot}
