% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{is_metanet}
\alias{is_metanet}
\alias{is.metanet}
\title{Is this object an metanet graph?}
\usage{
is_metanet(go)
}
\arguments{
\item{go}{igraph object}
}
\value{
logical
}
\description{
Is this object an metanet graph?
}
\examples{
data(c_net)
is_metanet(co_net)
}
