% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{links_stat}
\alias{links_stat}
\title{Link summary of the network}
\usage{
links_stat(
  go,
  group = "v_class",
  inter = "all",
  topN = 6,
  direct = T,
  legend_number = F,
  legend = T,
  legend_cex = 1,
  legend_position = c(left_leg_x = -1.6, left_leg_y = 1, right_leg_x = 1.2, right_leg_y =
    1),
  col_legend_order = NULL,
  group_legend_title = NULL,
  group_legend_order = NULL
)
}
\arguments{
\item{go}{igraph}

\item{group}{summary which group of vertex attribution in names(vertex_attr(go))}

\item{inter}{"positive", "negative", "all"}

\item{topN}{topN of group, default:5}
}
\value{
plot
}
\description{
Link summary of the network
}
\examples{
links_stat(co_net,topN=10)
modu_dect(co_net) -> co_net_modu
links_stat(co_net_modu,group="module")

c_net_plot(co_net,g_lay_polyarc(co_net,"v_class"))

}
