% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N-cycle.R
\name{load_N_data}
\alias{load_N_data}
\title{Load N-cycle data}
\usage{
load_N_data()
}
\value{
list
}
\description{
Load N-cycle data
}
\references{
\enumerate{
\item Tu, Q., Lin, L., Cheng, L., Deng, Y. & He, Z. NCycDB: a curated integrative database for fast and accurate metagenomic profiling of nitrogen cycling genes. Bioinformatics 35, 1040–1048 (2019).
\item Kuypers, M. M. M., Marchant, H. K. & Kartal, B. The microbial nitrogen-cycling network. Nat Rev Microbiol 16, 263–276 (2018).
}
}
