% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaNet-package.R
\docType{package}
\name{MetaNet-package}
\alias{MetaNet}
\alias{MetaNet-package}
\title{MetaNet: Network Analysis for Omics Data}
\description{
Comprehensive network analysis package. Calculate correlation network fastly, accelerate lots of analysis by parallel computing. Support for multi-omics data, search sub-nets fluently. Handle bigger data, more than 10,000 nodes in each omics. Offer various layout method for multi-omics network and some interfaces to other software (Gephi, Cytoscape, ggplot), easy to visualize. Provide comprehensive topology indexes calculation，including ecological network stability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/MetaNet}
  \item Report bugs at \url{https://github.com/Asa12138/MetaNet/issues}
}

}
\author{
\strong{Maintainer}: Peng Chen \email{pengchen2001@zju.edu.cn}

}
\keyword{internal}
