% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modu_dect}
\alias{modu_dect}
\title{Detect the modules}
\usage{
modu_dect(go, method = "cluster_fast_greedy", n_modu = 0)
}
\arguments{
\item{go}{a igraph object}

\item{method}{cluster_method:("cluster_walktrap","cluster_edge_betweenness","cluster_fast_greedy","cluster_spinglass")}

\item{n_modu}{transfer the modules less than n_modu to "others"}
}
\value{
a igraph object
}
\description{
Detect the modules
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
}
