% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modu_net}
\alias{modu_net}
\title{Generate a n-modules network}
\usage{
modu_net(
  module_number = 3,
  n_node_in_module = 30,
  intra_modu_density = 0.3,
  inter_modu_density = 0.01
)
}
\arguments{
\item{module_number}{number of modules}

\item{n_node_in_module}{number of nodes in each modules}

\item{intra_modu_density}{intra_modu_density, recommend bigger than 20*inter_modu_density, default:0.3}

\item{inter_modu_density}{inter_modu_density, default:0.01}
}
\value{
n-modules metanet
}
\description{
this is just a random generation method, the module number of result is not exactly the module_number, you can change the inter_modu_density and intra_modu_density to get the proper result.
}
\examples{
g1 <- modu_net()
get_n(g1)
plot(g1, mark_module = TRUE)
plot(g1, coors = g_lay(g1, zoom2 = 20))
plot(g1, coors = g_lay_polyarc(g1, group = "module"))
plot(g1, coors = g_lay_polygon(g1, group = "module"))
}
