% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modu_plot}
\alias{modu_plot}
\title{Plot a community (modularity)}
\usage{
modu_plot(go, coors = NULL, n_modu = 1, ...)
}
\arguments{
\item{go}{a igraph object}

\item{n_modu}{>n_modu modules will be ploted}

\item{...}{additional}

\item{wc}{a community object or a list contain wc and go}
}
\value{
a plot
}
\description{
Plot a community (modularity)
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
modu_plot(co_net_modu, n_modu = 20)
}
