% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_eigen}
\alias{module_eigen}
\title{Calculate the eigenvalue of each module and correlation of nodes and eigenvalue (node_eigen_cor).}
\usage{
module_eigen(go_m, totu)
}
\arguments{
\item{go_m}{network}

\item{totu}{original abundance table}
}
\description{
Calculate the eigenvalue of each module and correlation of nodes and eigenvalue (node_eigen_cor).
}
