% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_expression}
\alias{module_expression}
\title{Plot the expression of each modules}
\usage{
module_expression(go_m, totu, group = NULL, cor = 0.6)
}
\arguments{
\item{go_m}{network}

\item{totu}{original abundance table used for module_eigen().}

\item{group}{group variable for totu}

\item{cor}{the threshold for node_eigen_cor}
}
\description{
Plot the expression of each modules
}
\examples{
module_eigen(co_net_modu,totu)->co_net_modu
module_expression(co_net_modu,totu)
}
