% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_net}
\alias{multi_net}
\title{MetaNet networks}
\format{
three co_nets
\describe{
\item{tbls}{contians otutable rawdata}
\item{metas}{contians metadata}
\item{otus}{contians taxomomy table}
}
}
\usage{
c_net
}
\description{
MetaNet co_nets
}
