% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{my_network_tree}
\alias{my_network_tree}
\title{My Network plot}
\usage{
my_network_tree(
  test,
  vertex_anno = NULL,
  vertex_group = "v_group",
  vertex_class = "v_class",
  vertex_size = "size",
  ...
)
}
\arguments{
\item{test}{a dataframe with hierarchical structure}

\item{vertex_anno}{vertex annotation table}

\item{vertex_group}{vertex_group}

\item{vertex_class}{vertex_class}

\item{vertex_size}{vertex_size}

\item{...}{look for parameters in \code{\link[MetaNet]{c_net_plot}}}
}
\value{
metanet
}
\description{
My Network plot
}
\examples{
\donttest{
data(otutab, package = "pcutils")
cbind(taxonomy, num = rowSums(otutab))[1:10, ] -> test
my_network_tree(test)
}
}
