% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{net_par}
\alias{net_par}
\alias{c_net_index}
\title{Calculate all indexes of a network}
\usage{
net_par(go, mode = c("v", "e", "n", "all"), fast = TRUE)

c_net_index(go, force = FALSE)
}
\arguments{
\item{go}{igraph or metanet}

\item{mode}{calculate what? c("v", "e", "n", "all")}

\item{fast}{less indexes for faster calculate ?}

\item{force}{replace exist net_par}
}
\value{
a 3-elements list
\item{n_index}{indexs of the whole network}
\item{v_index}{indexs of each vertex}
\item{e_index}{indexs of each edge}
}
\description{
Calculate all indexes of a network

Add topological indexes for a network
}
\examples{
igraph::make_graph("Walther") \%>\% net_par()
c_net_index(co_net)->co_net_with_par
}
