% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4.other_plot.R
\name{netD3plot}
\alias{netD3plot}
\title{plot use networkD3}
\usage{
netD3plot(go, v_class = "v_class", ...)
}
\arguments{
\item{go}{metanet}

\item{v_class}{which attributes use to be v_class}

\item{...}{see \code{\link[networkD3]{forceNetwork}}}
}
\value{
D3 plot
}
\description{
plot use networkD3
}
\examples{
data("c_net")
plot(co_net2)
if (requireNamespace("networkD3")) {
  netD3plot(co_net2)
}
}
\seealso{
Other plot: 
\code{\link{as.ggig}()},
\code{\link{c_net_plot}()},
\code{\link{input_cytoscape}()},
\code{\link{input_gephi}()},
\code{\link{olympic_rings_net}()},
\code{\link{plot.ggig}()},
\code{\link{twocol_edgelist}()},
\code{\link{venn_net}()}
}
\concept{plot}
