% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{olympic_rings_net}
\alias{olympic_rings_net}
\title{Plot olympic rings using network}
\usage{
olympic_rings_net()
}
\value{
network plot
}
\description{
Plot olympic rings using network
}
\examples{
olympic_rings_net()
}
\seealso{
Other plot: 
\code{\link{as.ggig}()},
\code{\link{c_net_plot}()},
\code{\link{input_gephi}()},
\code{\link{netD3plot}()},
\code{\link{plot.ggig}()},
\code{\link{twocol_edgelist}()},
\code{\link{venn_net}()}
}
\concept{plot}
