% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otutab}
\alias{otutab}
\title{Microbiota community data}
\format{
three co_nets
\describe{
\item{tbls}{contians otutable rawdata}
\item{metas}{contians metadata}
\item{otus}{contians taxomomy table}
}
}
\usage{
otutab
}
\description{
Microbiota community data
}
