% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{output_gephi}
\alias{output_gephi}
\title{Output a graphml file for Gephi}
\usage{
output_gephi(go, file)
}
\arguments{
\item{go}{igraph}

\item{file}{graphml file name}
}
\description{
Output a graphml file for Gephi
}
\examples{
output_gephi(co_net,"test")
}
