% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_cor}
\alias{par_cor}
\title{Parallel calculate spearman correlation for one t(otutab)}
\usage{
par_cor(
  totu,
  totu2 = NULL,
  threads = 2,
  method = "spearman",
  p.adjust.methods = NULL
)
}
\arguments{
\item{totu}{t(otutab)}

\item{totu2}{t(otutab) or NULL}

\item{threads}{parallel mode threads, default 2}

\item{method}{spearman or not}

\item{p.adjust.methods}{how to adjust p-value (default:NULL, e.g.fdr)}
}
\value{
a list with 2 elements:
\item{r}{spearman correlation}
\item{p.value}{p.adjust.method = 'NULL}
}
\description{
Parallel calculate spearman correlation for one t(otutab)
}
\examples{
data("otutab")
t(otutab[1:100, ]) -> totu
par_cor(totu,4) -> corr

}
