% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_sim}
\alias{par_sim}
\title{Calulate distance for otutab}
\usage{
par_sim(totu, method = "bray")
}
\arguments{
\item{method}{Dissimilarity index, partial match to "bray", "euclidean"...}

\item{otutab}{an otutab data.frame, samples are columns, taxs are rows.}

\item{spe_nwk}{a phylo tree}
}
\value{
dist
}
\description{
Calulate distance for otutab
}
\seealso{
\code{\link[vegan]{vegdist};\link[picante]{unifrac}}
}
