% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_sparcc}
\alias{par_sparcc}
\title{SparCC correlation for a otutab}
\usage{
par_sparcc(totu, threads = 4, bootstrap = 100)
}
\arguments{
\item{totu}{an t(otutab)}

\item{threads}{default: 4}

\item{bootstrap}{default:100, recommend not less than 100}
}
\value{
list contain a correlation matrix and a bootstrap p_value matrix
}
\description{
SparCC correlation for a otutab
}
\examples{
par_sparcc(totu[,1:50])->sparcc_corr
}
