% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.net_compare.R
\name{plot.metanet_compare}
\alias{plot.metanet_compare}
\title{Plot a metanet_compare}
\usage{
\method{plot}{metanet_compare}(x, ...)
}
\arguments{
\item{x}{metanet_compare object}

\item{...}{add}
}
\value{
plot
}
\description{
Plot a metanet_compare
}
