% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3.RMT.R
\name{plot.rmt_res}
\alias{plot.rmt_res}
\title{Plot a rmt_res}
\usage{
\method{plot}{rmt_res}(x, ...)
}
\arguments{
\item{x}{rmt_res}

\item{...}{Additional arguments}
}
\value{
ggplot
}
\description{
Plot a rmt_res
}
