% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.calculate.R
\name{plot_corr_heatmap}
\alias{plot_corr_heatmap}
\title{Plot a correlation heatmap}
\usage{
plot_corr_heatmap(
  corr,
  show.p = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  ...
)
}
\arguments{
\item{corr}{a corr object, must contain 'r' and 'p.value' matrices.}

\item{show.p}{whether to show p-values as significance stars on the heatmap.}

\item{cluster_rows}{whether to cluster rows.}

\item{cluster_cols}{whether to cluster columns.}

\item{...}{additional arguments passed to \code{pheatmap::pheatmap}.}
}
\value{
plot of the correlation heatmap.
}
\description{
Plot a correlation heatmap
}
