% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2.manipulate.R
\name{plot_e_type_bar}
\alias{plot_e_type_bar}
\title{Plot e_type bar}
\usage{
plot_e_type_bar(go, mode = c("left", "right")[1], degree_threshold = 0)
}
\arguments{
\item{go}{metanet object}

\item{mode}{"left" or "right"}

\item{degree_threshold}{degree threshold}
}
\value{
ggplot
}
\description{
Plot e_type bar
}
\examples{
data("c_net")
plot_e_type_bar(co_net, degree_threshold = 10)
}
