% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.net_compare.R
\name{plot_multi_nets}
\alias{plot_multi_nets}
\title{Batch drawing multiple network diagrams}
\usage{
plot_multi_nets(graph_ls, nrow = NULL, ncol = NULL, multi_params_list = NULL)
}
\arguments{
\item{graph_ls}{a list containing igraph objects}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{multi_params_list}{a list of parameters for each network}
}
\value{
No value
}
\description{
Batch drawing multiple network diagrams
}
\examples{
plot_multi_nets(list(co_net, co_net2),
  multi_params_list = list(
    list(vertex.color = "skyblue"),
    list(vertex.color = "green3")
  )
)
}
