% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.topological.R
\name{plot_net_degree}
\alias{plot_net_degree}
\title{Plot degree distribution of networks}
\usage{
plot_net_degree(gols, net_names = NULL)
}
\arguments{
\item{gols}{list of metanet}

\item{net_names}{names of networks}
}
\value{
ggplot
}
\description{
Plot degree distribution of networks
}
