% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cohesion}
\alias{print.cohesion}
\title{Print method for 'cohesion' objects}
\usage{
\method{print}{cohesion}(x, ...)
}
\arguments{
\item{x}{'cohesion' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'cohesion' objects
}
