% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.coors}
\alias{print.coors}
\title{Print method for 'coors' objects}
\usage{
\method{print}{coors}(x, ...)
}
\arguments{
\item{x}{'coors' object}

\item{...}{additional arguments}
}
\description{
Print method for 'coors' objects
}
