% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{print.corr}
\alias{print.corr}
\title{Print method for 'corr' objects}
\usage{
\method{print}{corr}(x, ...)
}
\arguments{
\item{x}{'corr' object}

\item{...}{additional arguments}
}
\value{
No value
}
\description{
Print method for 'corr' objects
}
