% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ggig}
\alias{print.ggig}
\title{Print method for 'ggig' objects}
\usage{
\method{print}{ggig}(x, ...)
}
\arguments{
\item{x}{'ggig' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'ggig' objects
}
