% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{print.metanet}
\alias{print.metanet}
\title{Print method for 'metanet' objects}
\usage{
\method{print}{metanet}(x, ...)
}
\arguments{
\item{x}{'metanet' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'metanet' objects
}
