% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.metanet_compare}
\alias{print.metanet_compare}
\title{Print method for 'metanet_compare' objects}
\usage{
\method{print}{metanet_compare}(x, ...)
}
\arguments{
\item{x}{'metanet_compare' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'metanet_compare' objects
}
