% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robust}
\alias{print.robust}
\title{Print method for 'robust' objects}
\usage{
\method{print}{robust}(x, ...)
}
\arguments{
\item{x}{'robust' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'robust' objects
}
