% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robustness}
\alias{print.robustness}
\title{Print method for 'robustness' objects}
\usage{
\method{print}{robustness}(x, ...)
}
\arguments{
\item{x}{'robustness' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'robustness' objects
}
