% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.vulnerability}
\alias{print.vulnerability}
\title{Print method for 'vulnerability' objects}
\usage{
\method{print}{vulnerability}(x, ...)
}
\arguments{
\item{x}{'vulnerability' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print method for 'vulnerability' objects
}
