% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{rand_net_par}
\alias{rand_net_par}
\title{Net_pars of many random network}
\usage{
rand_net_par(go, reps = 99, threads = 4)
}
\arguments{
\item{go}{igraph}

\item{reps}{simulation time}

\item{threads}{threads}
}
\description{
Net_pars of many random network
}
