% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.calculate.R
\name{read_corr}
\alias{read_corr}
\title{Read a corr object}
\usage{
read_corr(filename)
}
\arguments{
\item{filename}{filename of .corr}
}
\value{
a corr object
}
\description{
Read a corr object
}
\seealso{
Other calculate: 
\code{\link{c_net_calculate}()},
\code{\link{cal_sim}()},
\code{\link{fast_cor}()},
\code{\link{p.adjust.table}()}
}
\concept{calculate}
