% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{robust_test}
\alias{robust_test}
\alias{plot.robustness}
\title{robust_test for a network}
\usage{
robust_test(go_ls, partial = 0.5, step = 10, reps = 9, threads = 1)

\method{plot}{robustness}(
  robust_res,
  indexs = c("nat_connectivity", "ave_path_len", "ave_degree"),
  mode = 1,
  ...
)
}
\arguments{
\item{partial}{how much percent vertexes be removed in total}

\item{step}{how many nodes be removed each time?}

\item{reps}{simulation number}

\item{threads}{threads}

\item{robust_res}{robust_test() result (robustness class)}

\item{indexs}{indexs selected to show}

\item{mode}{plot mode}

\item{go}{a igraph object}
}
\value{
dataframe(robustness class)

a ggplot
}
\description{
robust_test for a network

Plot robustness
}
\examples{
data("c_net")
robust_test(co_net,step=4)->robust_res
class(robust_res)
plot(robust_res,index="ave_degree",mode=2)
}
