% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{robustness}
\alias{robustness}
\title{Robustness after remove 50\% nodes or some hubs}
\usage{
robustness(go, keystone = F, reps = 9, threads = 4)
}
\arguments{
\item{go}{igraph}

\item{keystone}{remove 70\% keystone (default:False)}

\item{reps}{simulation time}

\item{threads}{threads}
}
\description{
Robustness after remove 50\% nodes or some hubs
}
\examples{
robustness(co_net)
modu_dect(co_net) -> co_net_modu
zp_analyse(co_net_modu,mode = 2)->co_net_modu
robustness(co_net_modu,keystone=T)
}
