% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{robustness_test}
\alias{robustness_test}
\alias{plot.robustness}
\title{Robustness_test for a network}
\usage{
robustness_test(go, partial = 0.5, step = 10, reps = 9, threads = 4)

\method{plot}{robustness}(
  robust_res,
  indexs = c("nat_connectivity", "ave_path_len", "ave_degree"),
  mode = 1
)
}
\arguments{
\item{go}{a igraph object}

\item{partial}{how much percent vertexes be removed in total}

\item{step}{how many nodes be removed each time?}

\item{reps}{simulation number}

\item{threads}{threads}

\item{robust_res}{robustness_test() result (robustness class)}

\item{indexs}{indexs selected to show}

\item{mode}{plot mode}
}
\value{
dataframe(robustness class)

a ggplot
}
\description{
Robustness_test for a network

Plot robustness
}
\examples{
data("c_net")
robustness_test(co_net,step=4)->robust_res
class(robust_res)
plot(robust_res,index="ave_degree",mode=2)
}
