% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.calculate.R
\name{save_corr}
\alias{save_corr}
\title{Save a corr object}
\usage{
save_corr(corr, filename = "corr")
}
\arguments{
\item{corr}{a corr object}

\item{filename}{filename without extension, default: "corr"}
}
\value{
a .corr file
}
\description{
Save a corr object
}
