% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{smallworldness}
\alias{smallworldness}
\title{Calculate small-world coefficient}
\usage{
smallworldness(go, reps = 99, threads = 1, verbose = TRUE)
}
\arguments{
\item{go}{igraph or metanet}

\item{reps}{simulation time}

\item{threads}{threads}

\item{verbose}{verbose}
}
\value{
number
}
\description{
Calculate small-world coefficient
}
\examples{
\donttest{
#set reps at least 99 when you run.
smallworldness(co_net, reps=9)
}
}
