% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{summ_2col}
\alias{summ_2col}
\title{Summaries two columns information}
\usage{
summ_2col(df, from = 1, to = 2, count = 3, direct = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{from}{first column name or index}

\item{to}{second column name or index}

\item{count}{(optional) weight column, if no, each equal to 1}

\item{direct}{consider direct? default: FALSE}
}
\value{
data.frame
}
\description{
Summaries two columns information
}
\examples{
test <- data.frame(a = sample(letters[1:4], 10, replace = TRUE),
    b = sample(letters[1:4], 10, replace = TRUE))
summ_2col(test, direct = TRUE)
summ_2col(test, direct = FALSE)
summ_2col(test, direct = TRUE) \%>\% pcutils::my_circo()
}
