% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{summ_module}
\alias{summ_module}
\title{Summary module index}
\usage{
summ_module(go_m, var = "v_class", module = "module", ...)
}
\arguments{
\item{go_m}{module metanet}

\item{var}{variable name}

\item{module}{which column name is module. default: "module"}

\item{...}{add}
}
\description{
Summary module index
}
\examples{
data("c_net")
modu_dect(co_net,n_modu=30) -> co_net_modu
summ_module(co_net_modu,var="v_class",module="module")
summ_module(co_net_modu,var="Abundance",module="module")
}
