% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.corr}
\alias{summary.corr}
\title{Summary method for 'corr' objects}
\usage{
\method{summary}{corr}(object, ...)
}
\arguments{
\item{object}{'corr' object}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Summary method for 'corr' objects
}
