% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.modules.R
\name{summary_module}
\alias{summary_module}
\title{Summary module index}
\usage{
summary_module(go_m, var = "v_class", module = "module", ...)
}
\arguments{
\item{go_m}{module metanet}

\item{var}{variable name}

\item{module}{which column name is module. default: "module"}

\item{...}{add}
}
\value{
ggplot
}
\description{
Summary module index
}
\examples{
data("c_net")
module_detect(co_net, n_node_in_module = 30) -> co_net_modu
summary_module(co_net_modu, var = "v_class", module = "module")
summary_module(co_net_modu, var = "Abundance", module = "module")
}
\seealso{
Other module: 
\code{\link{filter_n_module}()},
\code{\link{get_community}()},
\code{\link{get_module_eigen}()},
\code{\link{get_module}()},
\code{\link{module_detect}()},
\code{\link{module_eigen}()},
\code{\link{module_net}()},
\code{\link{to_module_net}()},
\code{\link{zp_analyse}()}
}
\concept{module}
