% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{to.ggig}
\alias{to.ggig}
\title{Transfer a igraph object to a ggig}
\usage{
to.ggig(go, net_pars = NULL, coors = NULL)
}
\arguments{
\item{go}{igraph}

\item{net_pars}{net_par() result}

\item{coors}{coordinates for nodes,columns: name, X, Y}
}
\value{
ggig object
}
\description{
Transfer a igraph object to a ggig
}
\examples{
coors=layout_with_fr(co_net, niter=999,grid="nogrid")
coors<-data.frame(name = V(co_net)$name,X=coors[,1],Y=coors[,2])
to.ggig(co_net,coors)->ggig
plot(ggig)
}
