% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.modules.R
\name{to_module_net}
\alias{to_module_net}
\title{Transformation a network to a module network}
\usage{
to_module_net(go)
}
\arguments{
\item{go}{metanet}
}
\value{
metanet with modules
}
\description{
Transformation a network to a module network
}
\seealso{
Other module: 
\code{\link{filter_n_module}()},
\code{\link{get_community}()},
\code{\link{get_module_eigen}()},
\code{\link{get_module}()},
\code{\link{module_detect}()},
\code{\link{module_eigen}()},
\code{\link{module_net}()},
\code{\link{summary_module}()},
\code{\link{zp_analyse}()}
}
\concept{module}
