% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{twocol_edgelist}
\alias{twocol_edgelist}
\title{Quick build a metanet from two columns table}
\usage{
twocol_edgelist(edgelist)
}
\arguments{
\item{edgelist}{two columns table (no elements exist in two columns at same time)}
}
\value{
metanet
}
\description{
Quick build a metanet from two columns table
}
\examples{
twocol=data.frame("col1"=sample(letters,30,replace = TRUE),
       "col2"=sample(c("A","B"),30,replace = TRUE))
twocol_net=twocol_edgelist(twocol)
plot(twocol_net)
}
