% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{venn_net}
\alias{venn_net}
\title{Venn network}
\usage{
venn_net(tab)
}
\arguments{
\item{tab}{data.frame (row is elements, column is group), or a list (names is group, value is elements)}
}
\value{
plot
}
\description{
Venn network
}
\examples{
data(otutab, package = "pcutils")
tab <- otutab[400:485, 1:3]
venn_net(tab) -> v_net
plot(v_net)
}
