% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{vulnerability}
\alias{vulnerability}
\title{Vulnerability}
\usage{
vulnerability(go, threads = 1)
}
\arguments{
\item{go}{igraph}
}
\value{
a vector
}
\description{
\deqn{Vi=\frac{E-Ei}{E}}
E is the global efficiency and Ei is the global efficiency after the removal of the node i and its entire links.
}
