% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{zp_analyse}
\alias{zp_analyse}
\title{Zi-Pi calculate}
\usage{
zp_analyse(go, mode = 2)
}
\arguments{
\item{go}{igraph object after modu_dect()}

\item{mode}{use 7-group (mode=1) or 4-group (mode=2)}
}
\value{
igraph
}
\description{
Zi-Pi calculate
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
zp_analyse(co_net_modu)->co_net_modu
zp_plot(co_net_modu)
zp_plot(co_net_modu,mode=3)
}
\references{
1. Guimerà, R. & Amaral, L. Functional cartography of complex metabolic networks. (2005) doi:10.1038/nature03288.
}
