% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{zp_plot}
\alias{zp_plot}
\title{Zi-Pi plot of vertexes}
\usage{
zp_plot(go, label = T, mode = 1)
}
\arguments{
\item{go}{igraph object after zp_analyse()}

\item{label}{show label or not}

\item{mode}{1,2,3}
}
\value{
a ggplot object
}
\description{
Zi-Pi plot of vertexes
}
\seealso{
\link{zp_analyse}
}
