% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{anno_edge}
\alias{anno_edge}
\title{Use dataframe to annotate edges of a igraph}
\usage{
anno_edge(go, anno_tab)
}
\arguments{
\item{go}{metanet a igraph object}

\item{anno_tab}{a dataframe using to annotate (with rowname or a name column)}
}
\value{
a annotated igraph object
}
\description{
Use dataframe to annotate edges of a igraph
}
\examples{
data("c_net")
data("otutab",package="pcutils")
anno_edge(co_net, taxonomy)->a
}
