% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{anno_vertex}
\alias{anno_vertex}
\title{Use dataframe to annotate vertexes of a igraph}
\usage{
anno_vertex(go, anno_tab)
}
\arguments{
\item{go}{a igraph object}

\item{anno_tab}{a dataframe using to annotate (with rowname or a name column)}
}
\value{
a annotated igraph object
}
\description{
Use dataframe to annotate vertexes of a igraph
}
\examples{
data("c_net")
data("otutab")
anno_vertex(co_net, taxonomy)->a
}
