% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-1.layout.R
\name{as_circle_tree}
\alias{as_circle_tree}
\title{Layout as a circle_tree}
\usage{
as_circle_tree()
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a circle_tree
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_line}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
