% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_line}
\alias{as_line}
\title{Layout as a line}
\usage{
as_line(angle = 0)
}
\arguments{
\item{angle}{anticlockwise rotation angle}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a line
}
\examples{
as_line()(co_net)
c_net_plot(co_net, coors = as_line(pi / 2))
}
