% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_polycircle}
\alias{as_polycircle}
\title{Layout as a polycircle}
\usage{
as_polycircle(n = 2)
}
\arguments{
\item{n}{how many circles of this polycircle}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a polycircle
}
\examples{
as_polycircle()(co_net)
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_circle_tree}()},
\code{\link{as_line}()},
\code{\link{as_polyarc}()},
\code{\link{as_polygon}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
