% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as_polygon}
\alias{as_polygon}
\title{Layout as a polygon}
\usage{
as_polygon(n = 3, line_curved = 0.5)
}
\arguments{
\item{n}{how many edges of this polygon}

\item{line_curved}{line_curved 0~0.5}
}
\value{
A two-column matrix, each row giving the coordinates of a vertex, according to the ids of the vertex ids.
}
\description{
Layout as a polygon
}
\examples{
as_polygon()(co_net)
}
\seealso{
Other layout: 
\code{\link{as_arc}()},
\code{\link{as_circle_tree}()},
\code{\link{as_line}()},
\code{\link{as_polyarc}()},
\code{\link{as_polycircle}()},
\code{\link{c_net_layout}()}
}
\concept{layout}
