% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{c_net_cal}
\alias{c_net_cal}
\title{Calculate spearman correlation for one or two t(otutab)}
\usage{
c_net_cal(
  totu,
  totu2 = NULL,
  method = "spearman",
  filename = "occor",
  p.adjust.method = NULL
)
}
\arguments{
\item{totu}{t(otutab)}

\item{totu2}{t(otutab) or NULL}

\item{method}{spearman, pearson}

\item{filename}{the prefix of saved files}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}
}
\value{
a list with 2 elements:
\item{r}{default: spearman correlation}
\item{p.value}{}
\item{p.adjust}{default p.adjust.method = NULL}
}
\description{
Calculate spearman correlation for one or two t(otutab)
}
\examples{
data("otutab")
t(otutab) -> totu
c_net_cal(totu) -> corr
metadata[,3:10] -> env
c_net_cal(totu, env) -> corr2
}
