% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_filter}
\alias{c_net_filter}
\title{Filter a network according to some attributes}
\usage{
c_net_filter(go, ..., mode = "v")
}
\arguments{
\item{go}{metanet}

\item{...}{some attributes of vertex and edge}

\item{mode}{"v" or "e"}
}
\value{
igraph
}
\description{
Filter a network according to some attributes
}
\examples{
data("multi_net")
c_net_filter(multi1,v_group\%in\%c("omic1","omic2"))
}
