% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_filter}
\alias{c_net_filter}
\title{Filter a network according to some attributes}
\usage{
c_net_filter(go, ...)
}
\arguments{
\item{go}{igraph}

\item{...}{some attributes of vertex and edge}
}
\value{
igraph
}
\description{
Filter a network according to some attributes
}
\examples{

c_net_filter(go,v_group=c("omic1","omic2"),e_class="intra")
}
