% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_from_edgelist}
\alias{c_net_from_edgelist}
\title{Construct a network from edge_list dataframe}
\usage{
c_net_from_edgelist(
  edgelist,
  vertex = NULL,
  direct = FALSE,
  e_type = NULL,
  e_class = NULL
)
}
\arguments{
\item{edgelist}{first is source, second is target, others are annotation}

\item{vertex}{vertex metadata}

\item{direct}{logical}

\item{e_type}{set e_type}

\item{e_class}{set e_class}
}
\value{
metanet
}
\description{
Construct a network from edge_list dataframe
}
\examples{
data(edgelist)
edge_net <- c_net_from_edgelist(arc_count, vertex = arc_taxonomy)
c_net_plot(edge_net)
}
