% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1.build.R
\name{c_net_from_edgelist}
\alias{c_net_from_edgelist}
\title{Construct a network from edge_list dataframe}
\usage{
c_net_from_edgelist(
  edgelist,
  vertex_df = NULL,
  direct = FALSE,
  e_type = NULL,
  e_class = NULL
)
}
\arguments{
\item{edgelist}{first is source, second is target, others are annotation}

\item{vertex_df}{vertex metadata data.frame}

\item{direct}{logical}

\item{e_type}{set e_type}

\item{e_class}{set e_class}
}
\value{
metanet
}
\description{
Construct a network from edge_list dataframe
}
\examples{
data(edgelist)
edge_net <- c_net_from_edgelist(arc_count, vertex_df = arc_taxonomy)
edge_net <- c_net_set(edge_net, vertex_class = "Phylum", edge_width = "n")
c_net_plot(edge_net)
}
\seealso{
Other build: 
\code{\link{c_net_build}()},
\code{\link{c_net_set}()},
\code{\link{c_net_update}()},
\code{\link{multi_net_build}()}
}
\concept{build}
