% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_save}
\alias{c_net_save}
\title{Save network file}
\usage{
c_net_save(go, filename = "net", format = "data.frame")
}
\arguments{
\item{go}{metanet network}

\item{filename}{filename}

\item{format}{"data.frame","graphml"}
}
\value{
No value
}
\description{
Save network file
}
\seealso{
Other manipulate: 
\code{\link{anno_edge}()},
\code{\link{anno_vertex}()},
\code{\link{c_net_annotate}()},
\code{\link{c_net_filter}()},
\code{\link{c_net_union}()},
\code{\link{get_e}()},
\code{\link{get_n}()},
\code{\link{get_v}()},
\code{\link{is_metanet}()}
}
\concept{manipulate}
