% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_union}
\alias{c_net_union}
\title{Union two networks}
\usage{
c_net_union(go1, go2)
}
\arguments{
\item{go1}{metanet object}

\item{go2}{metanet object}
}
\value{
metanet
}
\description{
Union two networks
}
\examples{
data("c_net")
co_net_union <- c_net_union(co_net, co_net2)
c_net_plot(co_net_union)
}
\seealso{
Other manipulate: 
\code{\link{anno_edge}()},
\code{\link{anno_vertex}()},
\code{\link{c_net_annotate}()},
\code{\link{c_net_filter}()},
\code{\link{c_net_save}()},
\code{\link{get_e}()},
\code{\link{get_n}()},
\code{\link{get_v}()},
\code{\link{is_metanet}()}
}
\concept{manipulate}
