% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{c_net_update}
\alias{c_net_update}
\alias{as.metanet}
\title{Update a metanet object or transform igraph object to metanet object}
\usage{
c_net_update(
  go,
  node_break = 5,
  edge_break = 5,
  initialize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{go}{a metanet object or igraph object}

\item{node_break}{node_break if v_class is numeric, default: 5}

\item{edge_break}{edge_break if e_type is numeric, default: 5}

\item{initialize}{initialize?}

\item{verbose}{verbose?}
}
\value{
metanet
}
\description{
Update a metanet object or transform igraph object to metanet object
}
\seealso{
Other build: 
\code{\link{c_net_build}()},
\code{\link{c_net_from_edgelist}()},
\code{\link{c_net_set}()},
\code{\link{multi_net_build}()}
}
\concept{build}
