% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{check_tabs}
\alias{check_tabs}
\title{Check tables}
\usage{
check_tabs(...)
}
\arguments{
\item{...}{tables}
}
\value{
format tables
}
\description{
A short description...
}
\examples{
data("otutab", package = "pcutils")
check_tabs(otutab)
}
