% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{Cohesion}
\alias{Cohesion}
\title{Cohesion calculate}
\usage{
Cohesion(otutab, reps = 200, threads = 4, mycor = NULL)
}
\arguments{
\item{otutab}{otutab}

\item{reps}{iteration time}

\item{threads}{threads}

\item{mycor}{a correlation matrix you want to use, skip the null model build}
}
\value{
a list with two dataframe
}
\description{
Cohesion calculate
}
\examples{
Cohesion(otutab[1:50,])->a
pctax::stackplot(abs(t(a$Cohesion)),metadata,groupID = "Group")
a$Cohesion\%>\%transmute(`neg:pos`=neg/pos)\%>\%pcutils::group_box(.,"Group",metadata)
}
\references{
1. Herren, C. M. & McMahon, K. Cohesion: a method for quantifying the connectivity of microbial communities. (2017) doi:10.1038/ismej.2017.91.
}
