% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{combine_n_modu}
\alias{combine_n_modu}
\title{Combine or cut modules to modu_num}
\usage{
combine_n_modu(go_m, modu_num = 5)
}
\arguments{
\item{go_m}{network}

\item{modu_num}{number of modules}
}
\description{
Combine or cut modules to modu_num
}
