% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-4.other_plot.R
\name{df2net_tree}
\alias{df2net_tree}
\title{Transform a dataframe to a network edgelist.}
\usage{
df2net_tree(test, fun = sum)
}
\arguments{
\item{test}{df}

\item{fun}{default: sum}
}
\value{
metanet
}
\description{
Transform a dataframe to a network edgelist.
}
\examples{
data("otutab", package = "pcutils")
cbind(taxonomy, num = rowSums(otutab))[1:20, ] -> test
df2net_tree(test) -> ttt
plot(ttt)
if (requireNamespace("ggraph")) plot(ttt, coors = as_circle_tree())
}
