% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{extract_sub_net}
\alias{extract_sub_net}
\title{Extract sub-network from the whole network}
\usage{
extract_sub_net(
  a_net,
  otutab,
  threads = 1,
  save_net = FALSE,
  fast = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{a_net}{the whole network}

\item{otutab}{otutab, these columns will be extract}

\item{threads}{threads, default: 1}

\item{save_net}{should save these sub_nets? FALSE or a filename}

\item{fast}{less indexes for faster calculate ?}

\item{verbose}{verbose}
}
\value{
a dataframe contains all sub_net parameters
}
\description{
Extract sub-network from the whole network
}
\examples{
data(otutab, package = "pcutils")
extract_sub_net(co_net, otutab) -> sub_net_pars
}
