% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{f_cor}
\alias{f_cor}
\title{Fast correlation}
\usage{
f_cor(totu, totu2 = NULL, method = c("pearson", "spearman"))
}
\arguments{
\item{totu}{t(otutab)}

\item{totu2}{t(otutab) or NULL}

\item{method}{spearman or pearson}
}
\value{
a list with 2 elements:
\item{r}{default: spearman correlation}
\item{p.value}{default: p-value of spearman correlation}
}
\description{
Fast correlation
}
\examples{
data("otutab", package = "pcutils")
t(otutab[1:100, ]) -> totu
f_cor(totu, method = "spearman") -> corr
}
