% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{f_cor}
\alias{f_cor}
\alias{par_cor}
\title{Fast correlation}
\usage{
f_cor(totu, totu2 = NULL, method = c("pearson", "spearman"))

par_cor(totu, totu2 = NULL, threads = 2, method = c("spearman", "pearson"))
}
\arguments{
\item{totu}{t(otutab)}

\item{totu2}{t(otutab) or NULL}

\item{method}{spearman or pearson}
}
\value{
a list with 2 elements:
\item{r}{}
\item{p.value}{}
}
\description{
Fast correlation

Parallel calculate correlation
}
\examples{
data("otutab")
t(otutab[1:100, ]) -> totu
f_cor(totu,method="spearman") -> corr
}
