% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{filter_n_modu}
\alias{filter_n_modu}
\title{Filter some modules as others}
\usage{
filter_n_modu(go_m, n_modu = 0, keep_id = NULL, delete = F)
}
\arguments{
\item{go_m}{module igraph}

\item{n_modu}{transfer the modules less than n_modu to "others"}

\item{keep_id}{keep modules ids, will not be "others"}

\item{delete}{logical, delete others modules? default:F, the others module will be "others".}
}
\description{
Filter some modules as others
}
