% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.topological.R
\name{fit_power}
\alias{fit_power}
\title{Fit power-law distribution for an igraph}
\usage{
fit_power(go, p.value = FALSE)
}
\arguments{
\item{go}{igraph}

\item{p.value}{calculate p.value}
}
\value{
ggplot
}
\description{
Fit power-law distribution for an igraph
}
\examples{
fit_power(co_net)
}
\seealso{
Other topological: 
\code{\link{extract_sample_net}()},
\code{\link{get_group_skeleton}()},
\code{\link{links_stat}()},
\code{\link{nc}()},
\code{\link{net_par}()},
\code{\link{rand_net_par}()},
\code{\link{rand_net}()},
\code{\link{smallworldness}()}
}
\concept{topological}
