% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{get_group_skeleton}
\alias{get_group_skeleton}
\alias{skeleton_plot}
\title{Get skeleton network according to a group}
\usage{
get_group_skeleton(go, Group = "v_class", count = NULL, direct = F)

skeleton_plot(tmp_go, ...)
}
\arguments{
\item{go}{network}

\item{tmp_go}{skeleton}

\item{group}{vertex column name}
}
\value{
skeleton network
}
\description{
Get skeleton network according to a group

skeleton_plot
}
\examples{
get_group_skeleton(co_net)->ske_net
skeleton_plot(ske_net)
x=tkplot(ske_net)
da <- tkplot.getcoords(x)
tkplot.close(x)
}
