% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{get_group_skeleton}
\alias{get_group_skeleton}
\alias{skeleton_plot}
\title{Get skeleton network according to a group}
\usage{
get_group_skeleton(go, Group = "v_class", count = NULL, top_N = 8)

skeleton_plot(ske_net, ...)
}
\arguments{
\item{go}{network}

\item{Group}{vertex column name}

\item{count}{take which column count, default: NULL}

\item{top_N}{top_N}

\item{ske_net}{skeleton}

\item{...}{additional parameters for \code{\link[igraph]{igraph.plotting}}}
}
\value{
skeleton network
}
\description{
Get skeleton network according to a group

Skeleton plot
}
\examples{
get_group_skeleton(co_net)->ske_net
skeleton_plot(ske_net)
}
