% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{input_corr}
\alias{input_corr}
\title{Import corr from .csv file}
\usage{
input_corr(filename)
}
\arguments{
\item{filename}{filename of .corr}
}
\value{
a corr object
}
\description{
Import corr from .csv file
}
\seealso{
Other calculate: 
\code{\link{c_net_calculate}()},
\code{\link{cal_sim}()},
\code{\link{fast_cor}()},
\code{\link{p.adjust.table}()}
}
\concept{calculate}
