% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{input_gephi}
\alias{input_gephi}
\title{Input a graphml file exported by Gephi}
\usage{
input_gephi(file)
}
\arguments{
\item{file}{graphml file exported by Gephi}
}
\value{
list contains the igraph object and coordinates
}
\description{
Input a graphml file exported by Gephi
}
\examples{
input_gephi("test.graphml")->gephi
c_net_plot(gephi$go,gephi$coors)
}
