% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{links_stat}
\alias{links_stat}
\title{Link summary of the network}
\usage{
links_stat(
  go,
  group = "v_class",
  inter = "all",
  topN = 6,
  number = F,
  direct = T,
  col_legend_order = NULL,
  group_legend_title = NULL,
  group_legend_order = NULL
)
}
\arguments{
\item{go}{igraph}

\item{group}{summary which group of vertex attribution in names(vertex_attr(go))}

\item{inter}{"positive", "negative", "all"}

\item{topN}{topN of group, default:5}
}
\value{
plot
}
\description{
Link summary of the network
}
\examples{
links_stat(co_net,topN=10)
modu_dect(co_net) -> co_net_modu
links_stat(co_net_modu,group="module")

c_net_plot(co_net,g_lay_polyarc(co_net,"v_class"))

}
