% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{manual_lay}
\alias{manual_lay}
\title{Interactive layout}
\usage{
manual_lay(go, canvas.width = 600, canvas.height = 400)
}
\description{
Interactive layout
}
\examples{
make_graph("Zachary")\%>\%manual_lay
}
