% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modu_dect}
\alias{modu_dect}
\title{Detect the modules}
\usage{
modu_dect(
  go,
  method = "cluster_fast_greedy",
  n_node_in_module = 0,
  delete = FALSE
)
}
\arguments{
\item{go}{an igraph object}

\item{method}{cluster_method: "cluster_walktrap", "cluster_edge_betweenness", "cluster_fast_greedy", "cluster_spinglass"}

\item{n_node_in_module}{transfer the modules less than n_node_in_module to "others"}

\item{delete}{logical, delete others modules? default:FALSE, the others module will be "others".}
}
\value{
an igraph object
}
\description{
Detect the modules
}
\examples{
data("c_net")
modu_dect(co_net) -> co_net_modu
}
