% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modu_net}
\alias{modu_net}
\title{Generate a n-modules network}
\usage{
modu_net(
  n_modu = 3,
  n_node_in_modu = 30,
  intra_modu_density = 0.3,
  inter_modu_density = 0.01
)
}
\arguments{
\item{n_modu}{number of modules}

\item{n_node_in_modu}{number of nodes in each modules}

\item{intra_modu_density}{intra_modu_density, recommend bigger than 20*inter_modu_density, default:0.3}

\item{inter_modu_density}{inter_modu_density, default:0.01}
}
\value{
n-modules metanet
}
\description{
Generate a n-modules network
}
\examples{
g1=modu_net()
get_n(g1)
plot(g1)
plot(g1,coors=g_lay(g1,zoom2 = 20))
plot(g1,coors=g_lay_polyarc(g1,group = "module"))
plot(g1,coors=g_lay_polygon(g1,group = "module"))
}
