% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_eigen}
\alias{module_eigen}
\title{Calculate the eigenvalue of each module and correlation of nodes and eigenvalue (node_eigen_cor).}
\usage{
module_eigen(go_m, totu, cor_method = "spearman")
}
\arguments{
\item{go_m}{module metanet}

\item{totu}{original abundance table}

\item{cor_method}{"pearson", "kendall", "spearman"}
}
\description{
Calculate the eigenvalue of each module and correlation of nodes and eigenvalue (node_eigen_cor).
}
