% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{multi_net_build}
\alias{multi_net_build}
\title{Multi-omics network build}
\usage{
multi_net_build(
  ...,
  mode = "full",
  method = "spearman",
  filename = "multi-net",
  p.adjust.method = NULL
)
}
\arguments{
\item{...}{some omics abundance tables}

\item{mode}{"full"}

\item{method}{spearman, pearson}

\item{filename}{the prefix of saved files}

\item{p.adjust.method}{see \code{\link[stats]{p.adjust}}}
}
\value{
metanet
}
\description{
Multi-omics network build
}
\examples{
data("multi_test")
multi_net_build(micro,metab,transc)->multi1
multi1=c_net_set(multi1,micro_g,metab_g,transc_g,vertex_class = c("Phylum","kingdom","type"))
multi1=c_net_set(multi1,data.frame("Abundance1"=colSums(micro)),data.frame("Abundance2"=colSums(metab)),data.frame("Abundance3"=colSums(transc)),vertex_size =paste0("Abundance",1:3))
c_net_plot(multi1)
multi1=c_net_set(multi1,vertex_class = c("kingdom","type"))
c_net_plot(multi1)

}
