% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{nc}
\alias{nc}
\title{Calculate natural_connectivity}
\usage{
nc(p)
}
\arguments{
\item{p}{a igraph object}
}
\value{
natural_connectivity (numeric)
}
\description{
Calculate natural_connectivity
}
\examples{
make_ring(10) \%>\% nc()
}
\references{
\code{\link[ggClusterNet]{nc}}
}
