% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.topological.R
\name{nc}
\alias{nc}
\title{Calculate natural_connectivity}
\usage{
nc(p)
}
\arguments{
\item{p}{an igraph or metanet object}
}
\value{
natural_connectivity (numeric)
}
\description{
Calculate natural_connectivity
}
\examples{
igraph::make_ring(10) \%>\% nc()
}
\references{
\code{`nc` in `ggClusterNet`}
}
\seealso{
Other topological: 
\code{\link{extract_sample_net}()},
\code{\link{fit_power}()},
\code{\link{get_group_skeleton}()},
\code{\link{links_stat}()},
\code{\link{net_par}()},
\code{\link{rand_net}()},
\code{\link{rand_net_par}()},
\code{\link{smallworldness}()}
}
\concept{topological}
