% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{net_par}
\alias{net_par}
\alias{get_net_par}
\title{Calculate all indexes of a network}
\usage{
net_par(go, mode = c("v", "e", "n", "all"), fast = T)

get_net_par(go, force = F)
}
\arguments{
\item{go}{a igraph object}

\item{mode}{calculate what? c("v", "e", "n", "all")}

\item{fast}{less indexes for faster calculate ?}
}
\value{
a 3-elements list
\item{n_index}{indexs of the whole network}
\item{v_index}{indexs of each vertex}
\item{e_index}{indexs of each edge}
}
\description{
Calculate all indexes of a network
}
\examples{
make_graph("Walther") \%>\% net_par()
get_net_par(co_net)->co_net_with_par
}
