% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{netD3plot}
\alias{netD3plot}
\title{plot use networkD3}
\usage{
netD3plot(go, v_class = "v_class", ...)
}
\arguments{
\item{go}{metanet}

\item{v_class}{which attributes use to be v_class}

\item{...}{see \code{\link[forceNetwork]{networkD3}}}
}
\description{
plot use networkD3
}
\examples{
plot(co_net2)
netD3plot(co_net2)
}
