% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_sim}
\alias{par_sim}
\title{Calculate distance for otutab}
\usage{
par_sim(totu, method = "bray", norm = FALSE)
}
\arguments{
\item{totu}{t(otutab)}

\item{method}{Dissimilarity index, partial match to "bray", "euclidean"...}

\item{norm}{hellinger normalization in tax (default:TRUE)}
}
\value{
1-dist, similarity
}
\description{
Calculate distance for otutab
}
\examples{
data("otutab", package = "pcutils")
t(otutab) -> totu
par_sim(totu) -> sim_corr
}
\seealso{
\code{\link[vegan]{vegdist}}
}
