% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_sparcc}
\alias{par_sparcc}
\title{SparCC correlation for a otutab}
\usage{
par_sparcc(totu, threads = 1, bootstrap = 100, verbose = TRUE)
}
\arguments{
\item{totu}{an t(otutab)}

\item{threads}{default: 1}

\item{bootstrap}{default:100, recommend not less than 100}

\item{verbose}{verbose}
}
\value{
list contain a correlation matrix and a bootstrap p_value matrix
}
\description{
SparCC correlation for a otutab
}
\examples{
\donttest{
data("otutab", package = "pcutils")
t(otutab) -> totu
par_sparcc(totu[, 1:50], 4) -> sparcc_corr
}
}
