% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{par_sparcc}
\alias{par_sparcc}
\title{SparCC correlation for a otutab}
\usage{
par_sparcc(totu, filename = "sparcc", threads = 4)
}
\arguments{
\item{totu}{an t(otutab)}

\item{filename}{saved file name}

\item{threads}{default: 4}
}
\value{
list contain a correlation matrix and a bootstrap p_value matrix
}
\description{
SparCC correlation for a otutab
}
\examples{
par_sparcc(totu)->sparcc_corr
}
