% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.net_compare.R
\name{plot.metanet_compare}
\alias{plot.metanet_compare}
\title{Plot a metanet_compare}
\usage{
\method{plot}{metanet_compare}(x, coors_com = NULL, mains = NULL, ...)
}
\arguments{
\item{x}{metanet_compare object}

\item{coors_com}{coors object}

\item{mains}{a vector of two strings for the main titles of the two networks}

\item{...}{add}
}
\value{
plot
}
\description{
Plot a metanet_compare
}
