% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{rand_net}
\alias{rand_net}
\title{Degree distribution comparison with random network}
\usage{
rand_net(go = go, plot = TRUE)
}
\arguments{
\item{go}{igraph object}

\item{plot}{plot or not}
}
\value{
ggplot
}
\description{
Degree distribution comparison with random network
}
\examples{
rand_net(co_net)
}
\seealso{
Other topological: 
\code{\link{extract_sample_net}()},
\code{\link{fit_power}()},
\code{\link{get_group_skeleton}()},
\code{\link{links_stat}()},
\code{\link{nc}()},
\code{\link{net_par}()},
\code{\link{rand_net_par}()},
\code{\link{smallworldness}()}
}
\concept{topological}
