% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{rand_net}
\alias{rand_net}
\title{Degree distribution comparison with random network}
\usage{
rand_net(go = go)
}
\arguments{
\item{go}{igraph object}
}
\value{
ggplot
}
\description{
Degree distribution comparison with random network
}
\examples{
rand_net(co_net)
}
