% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{RMT_threshold}
\alias{RMT_threshold}
\alias{rmt}
\title{Get RMT threshold for a correlation matrix}
\usage{
RMT_threshold(
  occor.r,
  min_threshold = 0.5,
  max_threshold = 0.8,
  step = 0.02,
  gif = FALSE,
  verbose = FALSE
)

rmt(occor.r, min_threshold = 0.5, max_threshold = 0.85, step = 0.01)
}
\arguments{
\item{occor.r}{a correlation matrix}

\item{min_threshold}{min_threshold}

\item{max_threshold}{max_threshold}

\item{step}{step}

\item{gif}{render a .gif file?}

\item{verbose}{verbose}
}
\value{
a r-threshold
}
\description{
Get RMT threshold for a correlation matrix

Get RMT threshold for a correlation matrix roughly
}
\examples{
\donttest{
\dontrun{
t(otutab) -> totu
c_net_cal(totu) -> corr
rmt(corr)
#0.69
c_net_build(corr,r_thres=0.69) -> co_net_rmt
RMT_threshold(corr$r,gif=TRUE)->rmt_res
}}
}
\references{
\enumerate{
\item J. Zhou, Y. Deng, FALSE. Luo, Z. He, Q. Tu, X. Zhi, Functional Molecular Ecological Networks (2010), doi:10.1128/mBio.00169-10.
}

\enumerate{
\item \code{https://matstat.org/content_en/RMT/RMThreshold_Intro.pdf}
}
}
