% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{RMT_threshold}
\alias{RMT_threshold}
\title{Get RMT threshold for a correlation matrix}
\usage{
RMT_threshold(occor.r)
}
\arguments{
\item{occor.r}{a correlation matrix}
}
\value{
a r-threshold
}
\description{
Get RMT threshold for a correlation matrix
}
\examples{
t(otutab) -> totu
c_net_cal(totu) -> corr
RMT_threshold(corr$r)
#0.65
}
\references{
1.  J. Zhou, Y. Deng, F. Luo, Z. He, Q. Tu, X. Zhi, Functional Molecular Ecological Networks (2010), doi:10.1128/mBio.00169-10.
}
