% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{plot.robust}
\alias{plot.robust}
\title{Plot robust}
\usage{
\method{plot}{robust}(
  x,
  indexes = c("Natural_connectivity", "Average_degree"),
  use_ratio = FALSE,
  mode = 1,
  ...
)
}
\arguments{
\item{x}{\code{robust_test()} result (robust object)}

\item{indexes}{indexes selected to show}

\item{use_ratio}{use the delete nodes ratio rather than nodes number}

\item{mode}{plot mode, 1~3}

\item{...}{additional arguments for \code{\link[pcutils]{group_box}}}
}
\value{
a ggplot
}
\description{
Plot robust
}
