% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{plot.robustness}
\alias{plot.robustness}
\title{Plot robustness}
\usage{
\method{plot}{robustness}(x, indexes = "Node_number", ...)
}
\arguments{
\item{x}{\code{robustness()} result (robustness object)}

\item{indexes}{indexes selected to show}

\item{...}{additional arguments for \code{\link[pcutils]{group_box}}}
}
\value{
a ggplot
}
\description{
Plot robustness
}
