% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{smallworldness}
\alias{smallworldness}
\title{Calculate small-world coefficient}
\usage{
smallworldness(go, reps = 99, threads = 1)
}
\arguments{
\item{go}{igraph}
}
\description{
Calculate small-world coefficient
}
\examples{
smallworldness(co_net)
}
