% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{summ_2col}
\alias{summ_2col}
\title{Summaries two columns information}
\usage{
summ_2col(df, from = 1, to = 2, count = 3, direct = F)
}
\arguments{
\item{df}{dataframe}

\item{from}{first column name or index}

\item{to}{second column name or index}

\item{count}{(optional) weight column, if no, each equal to 1}

\item{direct}{consider direct? default: F}
}
\description{
Summaries two columns information
}
\examples{
test=data.frame(a=sample(letters[1:4],10,replace = T),b=sample(letters[1:4],10,replace = T))
summ_2col(test,direct=T)
summ_2col(test,direct=F)
summ_2col(test,direct=T)\%>\%my_sankey()
}
