% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{to.ggig}
\alias{to.ggig}
\title{Transfer a igraph object to a ggig}
\usage{
to.ggig(go, coors = NULL)
}
\arguments{
\item{go}{igraph}

\item{coors}{coordinates for nodes,columns: name, X, Y}

\item{net_par_res}{net_par() result}
}
\value{
ggig object
}
\description{
Transfer a igraph object to a ggig
}
\examples{
coors=c_net_lay(co_net)
to.ggig(co_net,coors=coors)->ggig
plot(ggig)
to.ggig(multi1,coors=c_net_lay(multi1))->ggig
plot(ggig,labels_num=0.3)

}
