% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{to_module_net}
\alias{to_module_net}
\title{Transformation a network to a module network}
\usage{
to_module_net(go)
}
\arguments{
\item{go}{metanet}
}
\value{
metanet with modules
}
\description{
Transformation a network to a module network
}
