% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{vulnerability}
\alias{vulnerability}
\alias{plot.vulnerability}
\title{Vulnerability}
\usage{
vulnerability(go_ls, threads = 1)

\method{plot}{vulnerability}(x, ...)
}
\arguments{
\item{go_ls}{a igraph object or igraph list.}

\item{threads}{threads}

\item{x}{\code{vulnerability()} result (vulnerability object)}

\item{...}{add}
}
\value{
a vector

a ggplot
}
\description{
\deqn{Vi=\frac{E-Ei}{E}}
E is the global efficiency and Ei is the global efficiency after the removal of the node i and its entire links.
}
\examples{
\donttest{
\dontrun{
vulnerability(co_net)
}}
}
