% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.stability.R
\name{plot.vulnerability}
\alias{plot.vulnerability}
\title{Plot vulnerability}
\usage{
\method{plot}{vulnerability}(x, ...)
}
\arguments{
\item{x}{\code{vulnerability()} result (vulnerability object)}

\item{...}{add}
}
\value{
a ggplot
}
\description{
Plot vulnerability
}
