% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.modules.R
\name{zp_analyse}
\alias{zp_analyse}
\alias{zp_plot}
\title{Zi-Pi calculate}
\usage{
zp_analyse(go_m, mode = 2, use_origin = TRUE)

zp_plot(go, label = TRUE, mode = 1)
}
\arguments{
\item{go_m}{igraph object after \code{module_detect()}}

\item{mode}{plot style, 1~3}

\item{use_origin}{use original_module, default:TRUE, if FALSE, use module}

\item{go}{igraph object after zp_analyse()}

\item{label}{show label or not}
}
\value{
igraph

a ggplot object
}
\description{
Zi-Pi calculate

Zi-Pi plot of vertexes
}
\examples{
data("c_net")
module_detect(co_net) -> co_net_modu
zp_analyse(co_net_modu) -> co_net_modu
zp_plot(co_net_modu)
zp_plot(co_net_modu, mode = 3)

}
\references{
\enumerate{
\item Guimerà, R. & Amaral, L. Functional cartography of complex metabolic networks. (2005) doi:10.1038/nature03288.
}
}
\seealso{
Other module: 
\code{\link{filter_n_module}()},
\code{\link{get_community}()},
\code{\link{get_module_eigen}()},
\code{\link{get_module}()},
\code{\link{module_detect}()},
\code{\link{module_eigen}()},
\code{\link{module_net}()},
\code{\link{summary_module}()},
\code{\link{to_module_net}()}
}
\concept{module}
